+++ tests/test_mod_output.py
@@ -85,30 +85,30 @@ class TestSeeResultClass(unittest.TestCase):
         self.assertEqual(expected, filtered_result)
 
     @unittest.skip('Not implemented')
-    def test_filter_ignore_case_with_wildcard(self):
+    def test_filter_ignoring_case_with_wildcard(self):
         pass
 
-    def test_filter_ignore_case_with_regex_string(self):
+    def test_filter_ignoring_case_with_regex_string(self):
         obj = []
         pattern = '[AEIOU]{2}' # TODO: Change this to '/[AEIOU]{2}/'
         expected = ('.clear()', '.count()') if PY3 else ('.count()',)
 
-        filtered_result = see(obj).filter_ignore_case(pattern)
+        filtered_result = see(obj).filter_ignoring_case(pattern)
 
         self.assertIsInstance(filtered_result, output.SeeResult)
         self.assertEqual(expected, filtered_result)
 
     @unittest.skip('Not implemented')
-    def test_filter_exclude_with_wildcard(self):
+    def test_exclude_with_wildcard(self):
         pass
 
-    def test_filter_exclude_with_regex_string(self):
+    def test_exclude_with_regex_string(self):
         obj = []
         pattern = '[aeiou]{2}' # TODO: Change this to '/[aeiou]{2}/'
         excluded = set(('.clear()', '.count()')) if PY3 else set(('.count()',))
 
         unfiltered_result = see(obj)
-        filtered_result = see(obj).filter_exclude(pattern)
+        filtered_result = see(obj).exclude(pattern)
 
         self.assertIsInstance(filtered_result, output.SeeResult)
         self.assertEqual(set(unfiltered_result).difference(filtered_result),